refactor: extract logic from dealer animation
