+++ src/ui/modules/AnimationLab/AnimationLabPrimitives.ts
@@ -1,10 +1,30 @@
-import { Container, Sprite, Text, Texture } from "pixi.js";
+import { Container, Graphics, Sprite, Text, Texture } from "pixi.js";
 
 export const createCard = (tint: number, alpha = 1): Sprite => {
   const card = new Sprite(Texture.WHITE);
   card.tint = tint;
   card.alpha = alpha;
   return card;
+};
+
+export const createRoundedCardGraphic = (
+  width: number,
+  height: number,
+  radius: number,
+  fillColor: number,
+  strokeColor: number,
+  strokeAlpha: number,
+): Graphics => {
+  const graphic = new Graphics();
+  graphic
+    .roundRect(-width / 2, -height / 2, width, height, radius)
+    .fill(fillColor);
+  graphic.roundRect(-width / 2, -height / 2, width, height, radius).stroke({
+    color: strokeColor,
+    width: 3,
+    alpha: strokeAlpha,
+  });
+  return graphic;
 };
 
 export const clearContainer = (container: Container): void => {
